fix #1347

+++ src/network/Port.c
@@ -709,7 +709,7 @@ int swPort_http_static_handler(swHttpRequest *request, swConnection *conn)
         {
             date_format = SW_HTTP_ASCTIME_DATE;
         }
-        if (date_format && mktime(&tm3) - timezone >= file_mtime)
+        if (date_format && mktime(&tm3) - (int) timezone >= file_mtime)
         {
             response.length = response.info.len = snprintf(header_buffer, sizeof(header_buffer),
                     "HTTP/1.1 304 Not Modified\r\n"
